requirements.txt
pytesseract
pillow
flask
numpy
opencv-python
matplotlib